requirements.txt
keyboard==0.13.5
numpy==2.3.1
opencv_python==4.11.0.86
Pillow==11.3.0
pyautogui==0.9.54
